refactor: modularize api
