+++ AGENTS.md
@@ -14,6 +14,9 @@
 | App.tsx               | Главный компонент приложения, стартовая точка интерфейса            |
 | api.ts                | Функции для HTTP-запросов (авторизация, управление, CRUD)           |
 | api.test.ts           | Юнит-тесты для api.ts                                               |
+| modules/auth.ts         | Функции аутентификации пользователей |
+| modules/telegram.ts     | Интеграция Telegram (токены, поиск) |
+| modules/subscription.ts | Запросы и статистика подписок |
 | schema.prisma         | Схема базы данных Prisma                                            |
 | theme.css             | Основные стили интерфейса                                           |
 | PRD.md                | Продуктовые требования, описание сценариев использования            |
@@ -43,6 +46,10 @@
 - [ ] Добавить управление ролями пользователей (админ/юзер)
 - [ ] Настроить .gitignore для исключения node_modules, .env и сессионных файлов
 - [ ] Описывать каждый новый модуль/файл в таблице выше и в этом журнале
+
+### [2025-06-06]
+- [СДЕЛАНО] Вынесены функции из api.ts в модули auth.ts, telegram.ts и subscription.ts для лучшей модульности
+- [НОВАЯ ЗАДАЧА] Перенести остальные API-функции в отдельные файлы
 
 ---
