+++ api.ts
@@ -16,185 +16,10 @@ const PASSWORD_RESET_EXPIRY_HOURS = 24;
 // Promisify exec for async/await usage
 const exec = util.promisify(execCallback);
 
-// User Authentication
-export async function getCurrentUser() {
-  try {
-    const auth = await getAuth();
-    if (auth.status !== "authenticated") {
-      return null;
-    }
+export { getCurrentUser } from "./modules/auth";
+export { getUserSubscriptions, getSubscriptionStats } from "./modules/subscription";
+export { generateTelegramToken, verifyTelegramToken, getUserByTelegramId } from "./modules/telegram";
 
-    let user = await db.user.findUnique({
-      where: { id: auth.userId },
-    });
-
-    if (!user) {
-      // Create user if they don't exist yet
-      user = await db.user.create({
-        data: { id: auth.userId },
-      });
-    }
-
-    return user;
-  } catch (error) {
-    console.error("Error in getCurrentUser:", error);
-    throw new Error("Failed to get current user");
-  }
-}
-
-// Subscription Management
-export async function getUserSubscriptions() {
-  try {
-    const auth = await getAuth({ required: true });
-
-    return await db.subscription.findMany({
-      where: { userId: auth.userId },
-      orderBy: { endDate: "desc" },
-    });
-  } catch (error) {
-    console.error("Error in getUserSubscriptions:", error);
-    // Return empty array instead of throwing to handle gracefully in UI
-    return [];
-  }
-}
-
-export async function getSubscriptionStats() {
-  try {
-    const auth = await getAuth({ required: true });
-
-    const subscriptions = await db.subscription.findMany({
-      where: { userId: auth.userId },
-    });
-
-    const activeCount = subscriptions.filter(
-      (sub) => sub.status === "active",
-    ).length;
-    const expiredCount = subscriptions.filter(
-      (sub) => sub.status === "expired",
-    ).length;
-    const totalBandwidth = subscriptions.reduce(
-      (sum, sub) => sum + sub.bandwidth,
-      0,
-    );
-
-    return {
-      activeCount,
-      expiredCount,
-      totalBandwidth,
-      totalSubscriptions: subscriptions.length,
-    };
-  } catch (error) {
-    console.error("Error in getSubscriptionStats:", error);
-    // Return default stats object to handle gracefully in UI
-    return {
-      activeCount: 0,
-      expiredCount: 0,
-      totalBandwidth: 0,
-      totalSubscriptions: 0,
-    };
-  }
-}
-
-// Telegram Integration
-export async function generateTelegramToken() {
-  try {
-    const auth = await getAuth({ required: true });
-    const user = await db.user.findUnique({ where: { id: auth.userId } });
-
-    if (!user) {
-      throw new Error("User not found");
-    }
-
-    // Generate a secure token
-    const token = crypto.randomBytes(32).toString("hex");
-
-    // Set expiration to 24 hours from now
-    const expiresAt = new Date();
-    expiresAt.setHours(expiresAt.getHours() + 24);
-
-    // Create token in database
-    const telegramToken = await db.telegramToken.create({
-      data: {
-        token,
-        userId: user.id,
-        expiresAt,
-      },
-    });
-
-    return { token };
-  } catch (error) {
-    console.error("Error in generateTelegramToken:", error);
-    throw new Error("Failed to generate Telegram token");
-  }
-}
-
-export async function verifyTelegramToken(input: {
-  token: string;
-  telegramId: string;
-  username?: string;
-}) {
-  try {
-    // Find the token
-    const telegramToken = await db.telegramToken.findUnique({
-      where: { token: input.token },
-      include: { user: true },
-    });
-
-    if (!telegramToken) {
-      return { success: false, message: "Invalid token" };
-    }
-
-    if (telegramToken.isUsed) {
-      return { success: false, message: "Token already used" };
-    }
-
-    if (telegramToken.expiresAt < new Date()) {
-      return { success: false, message: "Token expired" };
-    }
-
-    if (!telegramToken.userId) {
-      return { success: false, message: "Token not associated with a user" };
-    }
-
-    // Update the user with Telegram info
-    await db.user.update({
-      where: { id: telegramToken.userId },
-      data: {
-        telegramId: input.telegramId,
-        telegramUsername: input.username,
-      },
-    });
-
-    // Mark token as used
-    await db.telegramToken.update({
-      where: { id: telegramToken.id },
-      data: { isUsed: true },
-    });
-
-    return { success: true, userId: telegramToken.userId };
-  } catch (error) {
-    console.error("Error in verifyTelegramToken:", error);
-    return { success: false, message: "Error verifying token" };
-  }
-}
-
-export async function getUserByTelegramId(input: { telegramId: string }) {
-  try {
-    const user = await db.user.findUnique({
-      where: { telegramId: input.telegramId },
-      include: { subscriptions: true },
-    });
-
-    if (!user) {
-      return null;
-    }
-
-    return user;
-  } catch (error) {
-    console.error("Error in getUserByTelegramId:", error);
-    return null;
-  }
-}
 
 // Subscription Plans
 export async function getSubscriptionPlans() {
